Revert "Mention `git tag` as an option for finding past versions."
This reverts commit f8673b05cabf9b7b43b3b04cbed2e7252c510304.

Authorea is junk. This is the second time I have caught that it has
erased my previous work by sneaking in deletions of past changes in
a new commit. This commit erased the 4 previous commits made to this
file. I reverted this commit, and in the merge conflict I kept the
mention of `git tag`.

The first commit I found this unwanted deletion of past work was
a9b8382:

https://github.com/jdblischak/git-for-science/commit/a9b8382b4e6a13553d10c95b0162320734c289c7

Conflicts:
	version-your-code.tex